Add duplicate = false flag on vm create
This adds an option called 'duplicate' on the vm module.
This makes the default action on VM create to *not* create a duplicate VM name even though hyper-v allows it.
It is possible to get the default hyper-v behaviour by having duplicate: true in the playbook.

If duplicate equals true then the module will always return changed: true